# sidecar-openapi.yaml
openapi: 3.0.3
info:
  title: FQNovel Signer Sidecar Internal API
  version: 0.1.0
  description: |
    Internal API contract for the Java signer/registerkey sidecar.

    Design assumptions:
    - Single user by default
    - Single Rust instance
    - Single Java sidecar instance
    - Single active upstream device profile

    This API is intended for local-only use via 127.0.0.1 or Unix Domain Socket.
servers:
  - url: http://127.0.0.1:18080
    description: Example local sidecar endpoint
tags:
  - name: signer
  - name: register-key
security:
  - InternalToken: []
paths:
  /internal/v1/sign:
    post:
      tags: [signer]
      summary: Generate upstream signature headers
      operationId: signRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: Signature generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignResponse"
              examples:
                signed:
                  value:
                    code: 0
                    message: success
                    data:
                      headers:
                        x-argus: "..."
                        x-gorgon: "..."
                        x-khronos: "1710000000"
                      signer_epoch: 12
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Signer temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unavailable:
                  value:
                    code: 1003
                    message: signer unavailable
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"

  /internal/v1/register-key/resolve:
    post:
      tags: [register-key]
      summary: Resolve the real AES key for a device profile and keyver
      operationId: resolveRegisterKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterKeyResolveRequest"
      responses:
        "200":
          description: Real key resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterKeyResolveResponse"
              examples:
                cache:
                  value:
                    code: 0
                    message: success
                    data:
                      device_fingerprint: "a8a8c5d4..."
                      keyver: 123
                      real_key_hex: "0123456789ABCDEF0123456789ABCDEF"
                      expires_at_ms: 1710003600000
                      source: cache
                refresh:
                  value:
                    code: 0
                    message: success
                    data:
                      device_fingerprint: "a8a8c5d4..."
                      keyver: 123
                      real_key_hex: "0123456789ABCDEF0123456789ABCDEF"
                      expires_at_ms: 1710003600000
                      source: refresh
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Register key version mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                mismatch:
                  value:
                    code: 1101
                    message: registerkey version mismatch
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"

  /internal/v1/register-key/invalidate:
    post:
      tags: [register-key]
      summary: Invalidate cached register key state for the current device fingerprint
      operationId: invalidateRegisterKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterKeyInvalidateRequest"
      responses:
        "200":
          description: Cache invalidated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterKeyInvalidateResponse"
              examples:
                invalidated:
                  value:
                    code: 0
                    message: success
                    data:
                      device_fingerprint: "a8a8c5d4..."
                      invalidated: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  /internal/v1/signer/reset:
    post:
      tags: [signer]
      summary: Request signer reset
      operationId: resetSigner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignerResetRequest"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignerResetResponse"
              examples:
                reset:
                  value:
                    code: 0
                    message: success
                    data:
                      accepted: true
                      signer_epoch: 13
                      cooldown_applied: false
                cooldown:
                  value:
                    code: 0
                    message: success
                    data:
                      accepted: true
                      signer_epoch: 13
                      cooldown_applied: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              value:
                code: 1001
                message: invalid request
                data: null
    Unauthorized:
      description: Missing or invalid internal token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                code: 1002
                message: unauthorized
                data: null
    InternalError:
      description: Internal sidecar error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                code: 1500
                message: internal sidecar error
                data: null

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
        server_time:
          type: integer
          format: int64
          nullable: true

    SignRequest:
      type: object
      required: [url, headers]
      properties:
        url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string

    SignResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/SignResult"
        server_time:
          type: integer
          format: int64
          nullable: true

    SignResult:
      type: object
      required: [headers, signer_epoch]
      properties:
        headers:
          type: object
          additionalProperties:
            type: string
        signer_epoch:
          type: integer
          format: int64

    RegisterKeyResolveRequest:
      type: object
      required: [device_profile]
      properties:
        device_profile:
          $ref: "#/components/schemas/DeviceProfile"
        required_keyver:
          type: integer
          format: int64
          nullable: true
          description: |
            When null, the sidecar may resolve the current/default key for the
            provided device profile.

    RegisterKeyResolveResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/RegisterKeyResolveResult"
        server_time:
          type: integer
          format: int64
          nullable: true

    RegisterKeyResolveResult:
      type: object
      required: [device_fingerprint, keyver, real_key_hex, expires_at_ms, source]
      properties:
        device_fingerprint:
          type: string
        keyver:
          type: integer
          format: int64
        real_key_hex:
          type: string
          pattern: "^[A-F0-9]{32}$"
        expires_at_ms:
          type: integer
          format: int64
        source:
          type: string
          enum: [cache, refresh]

    RegisterKeyInvalidateRequest:
      type: object
      required: [device_fingerprint]
      properties:
        device_fingerprint:
          type: string

    RegisterKeyInvalidateResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/RegisterKeyInvalidateResult"
        server_time:
          type: integer
          format: int64
          nullable: true

    RegisterKeyInvalidateResult:
      type: object
      required: [device_fingerprint, invalidated]
      properties:
        device_fingerprint:
          type: string
        invalidated:
          type: boolean

    SignerResetRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1

    SignerResetResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          enum: [success]
        data:
          $ref: "#/components/schemas/SignerResetResult"
        server_time:
          type: integer
          format: int64
          nullable: true

    SignerResetResult:
      type: object
      required: [accepted, signer_epoch, cooldown_applied]
      properties:
        accepted:
          type: boolean
        signer_epoch:
          type: integer
          format: int64
        cooldown_applied:
          type: boolean

    DeviceProfile:
      type: object
      required: [name, user_agent, cookie, device]
      properties:
        name:
          type: string
        user_agent:
          type: string
        cookie:
          type: string
        device:
          $ref: "#/components/schemas/UpstreamDevice"

    UpstreamDevice:
      type: object
      required:
        - aid
        - cdid
        - device_id
        - device_type
        - device_brand
        - install_id
        - resolution
        - dpi
        - rom_version
        - host_abi
        - update_version_code
        - version_code
        - version_name
        - os_version
        - os_api
      properties:
        aid:
          type: string
        cdid:
          type: string
        device_id:
          type: string
        device_type:
          type: string
        device_brand:
          type: string
        install_id:
          type: string
        resolution:
          type: string
        dpi:
          type: string
        rom_version:
          type: string
        host_abi:
          type: string
        update_version_code:
          type: string
        version_code:
          type: string
        version_name:
          type: string
        os_version:
          type: string
        os_api:
          type: string